requirements.txt
django
djangorestframework
django-filter
psycopg2-binary
pytest
pytest-django
pytest-cov
model-bakery
